pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crdb-dump"
version = "0.2.6"
description = "A CLI tool to export and import schema definitions and data from CockroachDB in SQL, JSON, YAML, or chunked CSV formats."

authors = [
    { name="Virag Tripathi", email="virag.tripathi@gmail.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent"
]
dependencies = [
    "sqlalchemy",
    "sqlalchemy-cockroachdb",
    "click",
    "pyyaml",
    "psycopg2-binary"
]

[project.scripts]
crdb-dump = "crdb_dump.cli:main"

[tool.setuptools.packages.find]
where = ["."]
